Add some docs, pull purchase params creation to extension function
