+++ feature/google/src/latestDependencies/java/com/revenuecat/purchases/google/billingClientParamBuilders.kt
@@ -3,8 +3,15 @@ package com.revenuecat.purchases.google
 import com.android.billingclient.api.BillingClient
 import com.android.billingclient.api.ProductDetails.SubscriptionOfferDetails
 import com.android.billingclient.api.QueryProductDetailsParams
+import com.android.billingclient.api.BillingFlowParams
 import com.android.billingclient.api.QueryPurchaseHistoryParams
 import com.android.billingclient.api.QueryPurchasesParams
+import com.revenuecat.purchases.common.errorLog
+import com.revenuecat.purchases.models.GooglePurchaseOption
+import com.revenuecat.purchases.models.GoogleStoreProduct
+import com.revenuecat.purchases.models.PurchaseOption
+import com.revenuecat.purchases.models.StoreProduct
+import com.revenuecat.purchases.models.googleProduct
 
 fun @receiver:BillingClient.ProductType String.buildQueryPurchaseHistoryParams(): QueryPurchaseHistoryParams? {
     return when (this) {
@@ -41,3 +48,27 @@ val SubscriptionOfferDetails.subscriptionBillingPeriod: String?
 
 val SubscriptionOfferDetails.isBasePlan: Boolean
     get() = this.pricingPhases.pricingPhaseList.size == 1
+
+fun StoreProduct.buildPurchaseParams(purchaseOption: PurchaseOption): BillingFlowParams? {
+    val googleProduct = this as? GoogleStoreProduct
+    if (googleProduct == null) {
+        errorLog("Product must be a GoogleStoreProduct.") //TODOBC5: Improve and move error message
+        return null
+    }
+
+    val googlePurchaseOption = purchaseOption as? GooglePurchaseOption
+    if (googlePurchaseOption == null) {
+        errorLog("PurchaseOption must be a GooglePurchaseOption.") //TODOBC5: Improve and move error message
+        return null
+    }
+
+    val token = purchaseOption.token
+    val productDetailsParamsList = BillingFlowParams.ProductDetailsParams.newBuilder().apply {
+        setOfferToken(token)
+        setProductDetails(googleProduct.productDetails)
+    }.build()
+
+    return BillingFlowParams.newBuilder()
+        .setProductDetailsParamsList(listOf(productDetailsParamsList))
+        .build()
+}
