+++ public/src/latestDependencies/java/com/revenuecat/purchases/models/GooglePurchaseOption.kt
@@ -8,9 +8,9 @@ data class GooglePurchaseOption(
     override val pricingPhases: List<PricingPhase>,
     override val storeProduct: StoreProduct,
     override val tags: List<String>,
+
     /**
-     * Token used to purchase. Only used for Google BC5 subscriptions, null otherwise.
+     * Token used to purchase
      */
-    val token: String? = null,
-) : PurchaseOption, Parcelable {
-}
+    val token: String,
+) : PurchaseOption, Parcelable
